// src/enums.ts
export enum EDataLayerType {
  BACKGROUND = "background",
  COLLIDERS = "colliders",
  SPAWNPOINTS = "spawnpoints",
  Object = "object",
}

export enum EColliderName {
  PLATFORM = "platform",
  EXIT = "exit",
  ENEMY = "enemy",
  SHOOTING_STAR = "shootingStar",
}

export enum EActions {
  UP = "up",
  DOWN = "down",
  LEFT = "left",
  RIGHT = "right",
  INHALE = "z",
  DOUBLE_JUMP = "x",
}

export enum ENPCState {
  IDLE = "idle",
  WALK = "walk",
  JUMP = "jump",
  LEFT = "left",
  RIGHT = "right",
}